Added SEOPress Post Meta transportation. SEOPress = massive spaghetti code.
Their massive spaghetti code took a month to churn through. SEOPress has the worst code I've come across, ever. Also found 18 security holes in their plugin, but they can suck it. Most of their features are duplicated, so it looks like it does more, but in fact, it does less and is being worse at it.