A massive list of proposed changes...
I started optimizing how we handle AJAX notices and remove the "Secure_Abstract" class altogether.

This class is part of our DRM and it's only making my life harder while not serving real customers better. It should go.

Alas, because it was so deeply intertwined with everything, removing its dependency created many issues. When I noticed them, I took a week off and forgot where I was working on -- as you can see, this commit is massive.

We need to park if for now since WP 6.7 broke this plugin on vital parts.